chore: rename HeaderButton into ButtonWithIcon
