+++ packages/legacy/core/App/components/buttons/HeaderHome.tsx
@@ -5,14 +5,14 @@ import { useTranslation } from 'react-i18next'
 import { Screens, TabStacks } from '../../types/navigators'
 import { testIdWithKey } from '../../utils/testable'
 
-import HeaderButton, { ButtonLocation } from './HeaderButton'
+import ButtonWithIcon, { ButtonLocation } from './ButtonWithIcon'
 
 const HeaderRightHome: React.FC = () => {
   const { t } = useTranslation()
   const navigation = useNavigation()
 
   return (
-    <HeaderButton
+    <ButtonWithIcon
       buttonLocation={ButtonLocation.Right}
       accessibilityLabel={t('Global.Home')}
       testID={testIdWithKey('HomeButton')}
